Rename folder to views
