+++ src/router/index.js
@@ -1,21 +1,21 @@
 import Vue from 'vue'
 import Router from 'vue-router'
-import Home from '@/components/Home'
-import About from '@/components/About'
-import AdminPoselka from '@/components/AdminPoselka'
-import BusGusevKgd from '@/components/BusGusevKgd'
-import BusGusevPoselok from '@/components/BusGusevPoselok'
-import BusKgdGusev from '@/components/BusKgdGusev'
-import BusPoselokGusev from '@/components/BusPoselokGusev'
-import Voda from '@/components/Voda'
-import KinoLumen from '@/components/KinoLumen'
-import KartaPoselka from '@/components/KartaPoselka'
-import TaxiVGuseve from '@/components/TaxiVGuseve'
-import TaxiVKaliningrade from '@/components/TaxiVKaliningrade'
-import PerevozkaDivana from '@/components/PerevozkaDivana'
-import GusevFok from '@/components/GusevFok'
-import Kebab from '@/components/Kebab'
-import FusionExpress from '@/components/FusionExpress'
+import Home from '@/views/Home'
+import About from '@/views/About'
+import AdminPoselka from '@/views/AdminPoselka'
+import BusGusevKgd from '@/views/BusGusevKgd'
+import BusGusevPoselok from '@/views/BusGusevPoselok'
+import BusKgdGusev from '@/views/BusKgdGusev'
+import BusPoselokGusev from '@/views/BusPoselokGusev'
+import Voda from '@/views/Voda'
+import KinoLumen from '@/views/KinoLumen'
+import KartaPoselka from '@/views/KartaPoselka'
+import TaxiVGuseve from '@/views/TaxiVGuseve'
+import TaxiVKaliningrade from '@/views/TaxiVKaliningrade'
+import PerevozkaDivana from '@/views/PerevozkaDivana'
+import GusevFok from '@/views/GusevFok'
+import Kebab from '@/views/Kebab'
+import FusionExpress from '@/views/FusionExpress'
 
 Vue.use(Router)
