+++ src/views/BusKgdGusev.vue
