+++ src/views/Home.vue
@@ -85,8 +85,7 @@
         </router-link>
       </div>
     </section>
-
-    </div>
+  </div>
 </template>
 
 <script>
